+++ Вопросы.txt
@@ -1,1 +1,1 @@
-1.
+1. Камера рывками догоняет игрока, когда он в режиме бега. Хорошо видно на оружии. Скорость движения камеры (множитель 7) больше скорости игрока (множитель 5). В режиме ходьбы (множитель 2) эффекта не видно.
